2.10.2
* fixed subcommands causing commands to show up in the list more than
once
+ added a way to show subcommands in the same type as the main command
* fixed the patreon api causing the bot to fail to update the list
- Removed First alias from Tags to make it just ListTags
+ added Tags subcommand to ListTags which just lists cc tags
* sped up the time it takes to send the special message when sail pins
something.
+ new GuildSetting CheckNewUsers (disabled by default) checks to see if
a user's account is less than 5 hours old and sends a message if it is.
+ new module JoinMessages, allows the creation and use of custom join
messages
+ new Channel setting Join_Channel
+ new commands: NewJoinMessage, EditJoinMessage, ListJoinMessages,
DeleteJoinMessage and JoinMessage

+++ src/main/java/com/github/vaerys/commands/pixels/TopUserForRole.java
@@ -50,7 +50,7 @@ public class TopUserForRole extends Command {
         List<Long> userIDs = command.guild.get().getUsersByRole(role).stream().map(IUser::getLongID).collect(Collectors.toList());
         if (userIDs.isEmpty()) {
             RequestHandler.deleteMessage(working);
-            return "> Couldn't find any users with that role!";
+            return "> Could not find any users with that role!";
         }
 
         userIDs.removeIf(f -> XpHandler.rank(command.guild.users, command.guild.get(), f) == -1);
@@ -60,6 +60,7 @@ public class TopUserForRole extends Command {
             long rank2 = XpHandler.rank(command.guild.users, command.guild.get(), o2);
             return Long.compare(rank1, rank2);
         });
+        if (userIDs.size() == 0) return "> Could not find any ranked users with that role.";
 
         if (index > userIDs.size()) {
             RequestHandler.deleteMessage(working);
@@ -89,7 +90,7 @@ public class TopUserForRole extends Command {
         int showing = (userIDs.size() > 5 ? 5 : userIDs.size());
 
         embed.withTitle("Top " + (userIDs.size() == 1 ? " user" : showing + " users") + " for role " + role.getName());
-        embed.withFooterText("Total users with this role: " + userIDs.size());
+        embed.withFooterText("Total ranked users with this role: " + userIDs.size());
         ProfileObject userProfile;
         UserObject userObject;
         NumberFormat nf = NumberFormat.getInstance();
